refactor: 简化 App.tsx 直接使用 MetricsBar 组件
- 移除 TopPane 中间层组件，减少组件层级
- 在 App.tsx 中直接配置和使用 MetricsBar
- 删除不再需要的 TopPane.tsx 文件
- 保持所有原有功能正常工作
- 添加第四阶段重构完成文档

这次重构提高了代码的可维护性，使组件结构更加清晰直观。
